collecting some app statistics and exposing through actuator; replaced js resources with min; new keys in app prod config

+++ eki-common/src/main/java/eki/common/data/Count.java
@@ -4,7 +4,7 @@ public class Count extends AbstractDataObject {
 
 	private static final long serialVersionUID = 1L;
 
-	private int value;
+	private long value;
 
 	public Count(int value) {
 		this.value = value;
@@ -22,7 +22,7 @@ public class Count extends AbstractDataObject {
 		this.value += byMuch;
 	}
 
-	public int getValue() {
+	public long getValue() {
 		return value;
 	}
 }
